Add hall owner verification email templates and update language files
- Introduced new email templates for hall owner verification and rejection notifications.
- Updated English and Arabic language files for hall availability, booking, and extra services with new keys and translations.
- Added error messages and status labels for hall availability management.
- Enhanced the extra service management with new validation messages and actions.

+++ app/Models/HallAvailability.php
@@ -329,6 +329,6 @@ class HallAvailability extends Model
      */
     public function getEffectivePrice(): float
     {
-        return $this->custom_price ?? $this->hall->getPriceForSlot($this->time_slot);
+        return (float) ($this->custom_price ?? $this->hall->getPriceForSlot($this->time_slot));
     }
 }
